+++ client/src/lib/util/fetchWithToken.ts
@@ -1,14 +1,18 @@
 import config from '../../config/config'
 
+export const getHeaders = (): Headers => {
+  const headers = new Headers()
+
+  headers.append('Authorization', `Bearer ${config.currentAccessToken}`)
+
+  return headers
+}
+
 export function fetchWithToken(uri: string): Promise<Response> {
   if (config.env.featureMyCollections && config.currentAccessToken) {
-    const headers = new Headers()
-
-    headers.append('Authorization', `Bearer ${config.currentAccessToken}`)
-
     return fetch(uri, {
       method: 'GET',
-      headers,
+      headers: getHeaders(),
     })
   }
   return fetch(uri)