+++ client/src/redux/api/mlMyCollectionsApi.tsx
@@ -2,8 +2,10 @@
 import { createApi } from '@reduxjs/toolkit/query/react'
 
 import { ICreateCollectionFormData } from '../../types/myCollections/ICreateCollectionFormData'
+import { IDeleteCollection } from '../../types/myCollections/IDeleteCollection'
 import { getDataApiBaseUrl } from '../../config/config'
 import { createCollectionObject } from '../../lib/myCollections/createCollection'
+import { getHeaders } from '../../lib/util/fetchWithToken'
 
 import { baseQuery } from './baseQuery'
 
@@ -22,9 +24,19 @@ export const mlMyCollectionsApi: any = createApi({
           defaultCollection,
         )
         return {
-          url: 'api/data/',
+          url: 'api/data',
           method: 'POST',
           body: collection,
+          headers: getHeaders(),
+        }
+      },
+    }),
+    deleteCollection: builder.mutation<any, IDeleteCollection>({
+      query: (collectionData) => {
+        const { type, id } = collectionData
+        return {
+          url: `api/data/${type}/${id}`,
+          method: 'DELETE',
         }
       },
     }),